GridTools.cpp splitted in two files: GridTools.cpp with general grid-related tools and GridTurbulence.cpp for turbulence-related functions. The refactroing is required for future extension of turbulence-related components, i.e. to introduce more realistic turbulent models. Otherwise, the GridTools.cpp would become too clutter.
