+++ test/testMagneticField.cpp
@@ -2,7 +2,7 @@
 
 #include "crpropa/magneticField/MagneticFieldGrid.h"
 #include "crpropa/Grid.h"
-#include "crpropa/GridTools.h"
+#include "crpropa/GridTurbulence.h"
 #include "crpropa/Units.h"
 #include "crpropa/Common.h"
